refactor(FarePolicy): enum과 decorator 분리
distance policy의 경우 연관성이 있어 분리함에 있어 고민필요
